+++ test/unit/server_test.js
@@ -9,6 +9,16 @@ describe("server.js tests", function () {
     this.axiosMock.restore();
   });
 
+  describe('Server.constructor', function () {
+    it("throws error for insecure server", function () {
+      expect(() => new StellarSdk.Server('http://horizon-live.stellar.org:1337')).to.throw(/Cannot connect to insecure horizon server/);
+    });
+
+    it("allow insecure server when opts.allowHttp flag is set", function () {
+      expect(() => new StellarSdk.Server('http://horizon-live.stellar.org:1337', {allowHttp: true})).to.not.throw();
+    });
+  });
+
   describe('Server._sendResourceRequest', function () {
 
     describe("requests all ledgers", function () {